+++ NEWS.md
@@ -3,7 +3,7 @@
 This document contains information on changes for each release.
 
 ## Unreleased changes
-- `create_question_random()` now has ability to turn off randomness check
+- `create_question_random()` now has more rigorous randomness check and ability to turn it off
 
 ## shinyquiz 0.0.1
 - Initial release